requirements.txt
numpy==1.22.3
pyserial==3.5
